fix: moved the models and auth service to the feature and made auth state service in core
