+++ dev-tools-frontend/src/app/core/guards/auth/auth-guard.ts
@@ -1,10 +1,10 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from '@core/services/auth/auth-service';
+import { AuthStateService } from '@core/services/auth-state-service';
 
 
 export const authGuard: CanActivateFn = () => {
-  const auth = inject(AuthService);
+  const auth = inject(AuthStateService);
   const router = inject(Router);
 
   return auth.isAuthenticated() ? true : router.parseUrl('/login');
