+++ dev-tools-frontend/src/app/core/services/auth-state-service.ts
@@ -1,25 +1,32 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable, inject, signal } from '@angular/core';
-import { Observable, tap } from 'rxjs';
-import { LoginRequest } from '@core/models/auth/loginRequest';
-
+import { Observable, catchError, map, of, tap } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
-export class AuthService {
+export class AuthStateService {
   private readonly http = inject(HttpClient);
 
   private readonly _isAuthenticated = signal(false);
   readonly isAuthenticated = this._isAuthenticated.asReadonly();
 
-  login(credentials: LoginRequest): Observable<void> {
-    return this.http
-      .post<void>('/login', credentials, { withCredentials: true })
-      .pipe(tap(() => this._isAuthenticated.set(true)));
+  markAuthenticated(): void {
+    this._isAuthenticated.set(true);
   }
 
   logout(): Observable<void> {
     return this.http
       .post<void>('/logout', {}, { withCredentials: true })
       .pipe(tap(() => this._isAuthenticated.set(false)));
+  }
+
+  checkSession(): Observable<void> {
+    return this.http.get<void>('/session', { withCredentials: true }).pipe(
+      tap(() => this._isAuthenticated.set(true)),
+      map(() => undefined),
+      catchError(() => {
+        this._isAuthenticated.set(false);
+        return of(undefined);
+      }),
+    );
   }
 }
